Enable Trace in shipping: TraceLog
Introduce a new configuration (UE_TRACE_ENABLED_SHIPPING_EXPERIMENTAL) which optionally enables select parts of the TraceLog library in shipping configurations. In this configuration, only a subset of trace events and logs are active. These needs to be de specificially for shipping using a new set of macros. Addtionally this change adds the ability to configure other parts of TraceLog functionality such as the socket based control layer, tracing to sockets or files and if important events are allowed. This is currently an experimental feature.

#rb Catalin.Dragoiu, ionut.matasaru

[CL 36033751 by johan berg in ue5-main branch]