Improve CookStatsFile to include a more complete set of metrics
To better track and understand cook performance for the content worker, we require a good bit more information that what the cook profile currently dumped out via -CookStatsFile=... provides. This change aims to make the metrics provided there more completel, such that they have parity with the stats we print to the log (but in a format that makes them easier to parse for tools).

The most noteworthy change is that we break the format slightly. Previously, all steps from the cook profile lived at the top-level of the output json object. With this change, they now sit in a subobject "CookProfile", so they do not pollute the other stats.

[RN] Cooks running with -CookStatsFile=... now produce a more comprehensive set of stats. The format has changed slightly, such that the previous members of the json object now live in a subobject called "CookProfile"
#rb dan.elksnitis, eric.knapik, Jason.Nadro

[CL 36186318 by sebastian schoner in 5.5 branch]