DaySequenceActor: Replicate playback by default
I think this is reasonable to do by default because a user can just make the actor replicate or not and it will yield expected behavior. In the current form the actor can replicate and it still might not be obvious how to synchronize playback. I think the vast majority of cases will have this set to true.

#ushell-cherrypick of 36128771 by nickolas.drake

#rb lonnie.li

[CL 36194958 by nickolas drake in 5.5 branch]