DaySequence: Fix crash when inspecting modifier subsequences in the root sequence while a level is being validated
I think this crash could probably occur in scenarios other than level validation. The basic issue is as follows:
1) A user is inspecting/focused on a modifier subsequence section
2) The modifier removes its subsequence section from the root sequence
3) An evaluation occurs

The result is a nullptr dereference because much of the core Sequencer code assumes the focused subsequence is always valid. The fix is to introduce a static delegate on ADaySequenceActor with the signature void(UMovieSceneSubSection*) that is broadcast by the modifier when its subsection is removed. The day sequence editor module monitors this delegate and changes the focused sequence to the root sequence if the currently focused subsequence is the same as the one assigned to the section passed as a parameter when invoking the delegate.

#rb lonnie.li
#jira UE-223968

[CL 36195040 by nickolas drake in 5.5 branch]