Cameras: fix collision node's safe position lookup
Always using the pivot and/or the owner evaluation context doesn't work well when the collision node is in the global stack. This change adds an option to specify where the safe position should be by default, and in particular allows getting it from the active context (from the main blend stack) or from the player pawn. There is also a new option to activate/deactivate collision in a smooth way.

Debug display has been improved, and potential log spam reduced.

#jira UE-223905

[CL 36199100 by ludovic chabant in 5.5 branch]