[ChaosVD] Fix issue where the Main tab and therefore all the widgets are leaked when CVD is closed
This is fallout from a change made last week. It happens because the CVD main tab passes itself to the custom detail instances. Although the instances itself keep a weak ptr to the main tab, the Delegate instance used to create the customization object instances, keeps a shared reference to the main tab.

The problem aside form the leak itself is that all CVD playback logic is destroyed at that point, which means any late delegate triggering ends up in a editor crash.

The fix is to pass the toolkit host (CVD Main Tab) as a weak ptr

#jira UE-224188
[FYI] Benn.Gallagher
#tests I tested that after the change, when CVD is closed the the destructors of the widgets for each tab are called as expected, and the editor crash does not happen anymore

[CL 36255757 by sergio gardeazabal in 5.5 branch]