MPCook: CookPackageSplitter: Fix for error "GenerationHelper for package %s is still allocated at end of cooksession...."
Fix a missing record that a generated package has completed its cook in some cases where its cook was suppressed. Suppressed packages were reported if suppressed on remote CookWorkers, or if suppressed when in the save state on the CookDirector, but if suppressed when in load state or earlier on the CookDirector they were not reported. Add code to report them as completed whenever their ParentGenerationHelper pointer is removed, which happens when they get demoted to idle.

Also add storage of the suppress cook reason so that diagnostics that later discover they were not cooked can report why they were not cooked.

#rnx
#rb Zousar.Shaker

[CL 36261010 by matt peters in 5.5 branch]