Cameras: more fixes for auto-resetting variables used from Blueprint
Multiple calls to SetValue from Blueprint would result in ever rising ref-counts for auto-resetting variables. This change keeps a separate flag for uses from BP/Verse/etc. The flag is cleared every frame, so when BP/Verse/etc. stops writing to the variable, it doesn't count as a reference anymore and the variable can be deleted if nothing else references it.

This change required doing a pass on the WrittenThisFrame flag:

- Clean-up how these flags are cleared in the camera system update.
- Correctly propagate the flag when overriding/interpolating variable tables.
- Option for not setting it inside low-level variable table API.

#jira UE-223952

[CL 36278794 by ludovic chabant in 5.5 branch]