Chaos Cloth Asset: fix Aerodynamics debug draw.
Softs Evolution doesn't calculate Aerodynamics forces in two passes, so the Forces buffer is not calculated. Instead just calculate the force based on the current state. This will represent the force for next frame instead of current frame, but most of our debug display shows based on current (end of frame) state rather than last frame, so this is consistent with that.

Also add a user-scale menu option and set the Simulation Statistics debug display to true, and fix a DPI scale bug in the text visualization.

#rb kriss.gossart
#jira UE-224203

[CL 36352223 by alex mcadams in 5.5 branch]