[ChaosVD] Speculative fix for a crash that can happen while playing back a recording with particles being individually destroyed before the entire solver is destroyed.
There is a null check missing on the scene outliner code. I already started a thread with the team owning it to see if it is something that is recoverable or should be an assert.

Meanwhile, this is a patch that will remove any invalid outliner items before we broadcast the update. As a workaround for UE-193877, CVD needs to enqueue all ADD/REMOVE events, de-duplicate them and process them later within a determined time budget. What seems to be happening is that we enqueue an add event, but then the underlying actor is destroyed before we process that event. To that to happen the solver that owns the particle needs to be unloaded right after, and the amount of updates to process in that frame has to be large enough to not fit within the 2ms budget, meaning the add event is processed a few frames later.

#jira UE-211836
[REVIEW]
#rb Benn.Gallagher

[CL 36360104 by sergio gardeazabal in 5.5 branch]