[PCG] GPU Static Mesh Spawner - support spawn by-attribute.
Carves out space for InputPointCount X PrimitiveCount instances for now (when using by-attribute spawning - weighted spawning remains the same). In the future dynamic analysis of the spawner input points could reduce this.

Requires string handling which isn't supported in HLSL. Represent using ints so they can still be read/selected/written in kernels.

String table is built at execution time which has all strings (or SOP paths) that enter the GPU graph. We need to know all strings that can arrive on the input pin of an SM spawner so we know what primitives to set up. Did a number of updates to data analysis/description to enable this.

Also added "dump data description" logging to GPU nodes for easy inspection of the data analysis.

#jira UE-217681
#rb julien.lheureux, wyatt.marvil
#rnx

[CL 36383358 by huw bowles in 5.5 branch]