[AVCodecs] Only encode VP9 video with an even dimension. This prevents crash in windows and linux.
Problem is internally in libvpx, valid_ref_frame_size checks the dimensions. When A p-frame is encoded, the UV size from a odd frame size causes this check to fail. Two function pointers for scaling are then not set. These function pointers then get called which causes the crash.

Set bExternalRefControl to true, which also prevents the crash on windows but not linux.

#rb William.Belcher
[FYI] Denis.Phoenix, Nazar.Rudenko, aidan.possemiers

[CL 36421881 by eden harris in 5.5 branch]