[AutoRTFM] Reduce chance of page faults when looking for "magic-mike" function prefix
A "magic-mike" prefix is a 64-bit integer placed before strongly-linked "open" functions, which holds the relative address to the closed variants. This is used as a size and performance optimization to avoid the cost of a function-map registration and lookup.

Weak-linked functions were not emitted with the prefix, as these many of these do not have a closed variant (standard library, third-party code, modules compiled without AutoRTFM, etc).

Because only some functions had the prefix, there was the possibility of a function's address landing on the start of a page, where the previous page was not readable. Attempting to load the "magic-mike" in this situation would cause a page fault.

We were specifically seeing this page fault happen in FunctionMapAdd(), called by the AutoRTFM function regisistration for a weak-linked function (inlined method) internal to a AutoRTFM compiled module.

This CL attempts to reduce the chance of this issue by emitting the "magic-mike" even for weak-linked functions. For the case where we had a non-public weak-linked function compiled inside an autortfm-enabled module, this should fix the page fault.

Note that weak-linked functions can be compiled by autortfm-enabled and autortfm-disabled (e.g. a template method in a shared header), so there's still a chance the no-prefix version could still picked by the linker and end up at the start of a page with a no-read page before it. However, this is less likely than what we have right now.

Also apply a fix to MallocLeakDetection.cpp from FNMain so that we can build the AutoRTFMTests.

Taken from Ben.Clayton change!

#rb Brandon.Schaefer
[FYI] Ben.Clayton

[CL 36447880 by brandon schaefer in 5.5 branch]