Sequencer: remove buggy support for weighted blendables in camera animations/shakes
The previous code was trying to save and restore any weighted blendables added by the Sequence onto the camera's PostProcessSettings, in order to combine them with any weighted blendables given by the game camera or some post-FX volume or other. The problem was that this code was also stomping the weight of pre-existing blendables, which prevented the original system that added them to change that weight.

This change removes this code altogether -- we have no support for weighted blendables in Sequencer, so the only way we could add weighted blendables in a camera animation or shake is if someone made a custom 3rd party track for it. Since this is highly unlikely, we are just removing this for now because it's not worth going through the trouble of finding the delta between 2 weighted arrays.

#udn 12002888
#rb Max.Chen

[CL 36461203 by ludovic chabant in 5.5 branch]