[PCG] Fix higen grid links losing data when upstream pin label is not 'Out'.
In a previous change (35826801), we needed to propagate pin type downstream across higen grid links, so that GPU compilation could access the pin type. This exposed issues with missing labels and we assigned virtual labels to the inputs and outputs of a grid link - 'In' and 'Out'. However this only worked when the upstream pin label is also 'Out'.

This change forwards the upstream pin properties all the way through the grid link and doesn't artificially mangle pin labels at any point of the link.

#jira UE-225011
#rb julien.lheureux, wyatt.marvil
#rnx

[CL 36512284 by huw bowles in 5.5 branch]