Overhaul content worker timing analytics
The content worker currently just emits the raw CookStats.json file, which is then ingested by our analytics backend. Previous CLs I put into the engine improved the coverage of these analytics.

This CL improves the analytics that we output by instrumenting major steps of the cook (times blocked on shader compilation, time spent in validation, time spent per asset type etc.), adding timings for packaging, staging, and chunking, and by adding generally useful numbers such as the number of assets and shaders built as well as DDC utilization.

A slight quirk here is that we need to get the timings from multiple UAT steps, which requires passing information back and forth through files. It's not pretty, but that's what we have.

#rnx

[CL 36746567 by sebastian schoner in 5.5 branch]