Post Process Alpha is preserved in SceneColor where possible for post process materials that don't write alpha.  This includes making it so materials that read and write from SceneColor, but don't write alpha, automatically read and write alpha.  For materials that only write to SceneColor, a Color Write mask is set to mask the alpha channel in the destination.  An important bug with UserSceneTextures was also fixed where a final pass of a post process chain wasn't properly identified if it wrote to a UserSceneTexture overridden to "SceneColor".
Propagated alpha is handled by storing the alpha to a local variable when PostProcessInput0 (SceneColor) is fetched from either a SceneTexture or UserSceneTexture node, then writing the fetched value to the output.  The propagation of alpha is free in any material which reads from a SceneTexture, but an extra scalar and vector ALU is generated per UserSceneTexture, since the alpha propagation is conditional based on whether a given input has SceneColor assigned as its input (based on the value of the UserSceneTextureSceneColorInput parameter).  The first UserSceneTexture vector ALU will be free, rolled into the output assignment.  The material can be set up to output opacity to optimize that minor overhead out completely if desired.

An advantage of this approach is that it even works with custom HLSL, since it just requires SceneTextureLookup to be reached by any means, and it also straightforwardly handles UserSceneTextures reading and writing SceneColor.  Some complexity is added in that SceneTexture or UserSceneTexture nodes feeding into custom HLSL, where the input pin value itself isn't used, now need to be explicitly dead stripped, as the side effect of writing to PropagatedAlpha prevents the shader compiler from optimizing out calls to SceneTextureLookup based on the return value being unused in downstream code.  See UMaterialExpressionCustom::Compile for where that is handled (with detailed comments).

#jira UE-133044
#rb eric.renaudhoude massimo.tristano jason.nadro

[CL 36747746 by jason hoerner in 5.5 branch]