DaySequenceModifierComponent: Add BP accessors to several properties
A number of our properties require non-trivial accessing logic (at least when setting, not always when getting). For example, setting DayNightCycle requires the modifier component to reenable and setting UserBlendWeight requires the set value be clamped.

Importantly note: we don't need to add accessors to properties that didn't already have them, we just need to make sure no accessors are BP callable. We can always introduce and remove accessors as long as they are invoked by BPs that are attempting to directly set the property because of UHT magic.

Another important note: I have observed that Setter/Getter markup is not respected by the details panel. I think most other things do respect it. For now I am just handling this in PostEditChangeProperty by routing the new value into the setter for properties that need this behavior. Trivial setters don't need this treatment.

This change also corrects the DayNightCycle setter to properly reenable the modifier on set.

#rb lonnie.li


#changelist validated
#virtualized

[CL 36753085 by nickolas drake in 5.5 branch]