Fixed animation budget allocator debug rendering sometimes getting removed
When ticking offscreen meshes (only possible with native code driving significance), sometimes an empty strting could be passed to debug rendering, removing all text for that actor. As we passed nullptr for the base actor, the world settings actor is used instead as the key, so passing an empty string cleared all debug text. The fix is to pass the correct base actor (and a zero vector for the location).

#jira UE-207751
#rb euan.carmichael, Jurre.deBaare

[CL 36754213 by thomas sarkanen in 5.5 branch]