[NDIMedia] Fix interoperability with NDI IO plugin by allowing NDI Media to use the system's NDI runtime instead of the bundled one.
The NDI Media plugin comes with NDI runtime bundled. However, it is an older version, and if users install SDK or Tools, they are likely to have a more recent version. For both plugins to work correctly they should have the same version. So we need to enable users to select which runtime they want to use.

The minimum version required to work is NDI 5 because the "NDIlib_v5_load" function is used. It works with NDI 6 +.

#rb Loris.Rossi

[CL 36755606 by jean perrier in 5.5 branch]