[PCG] Fix failed compute graph execution when there are upstream reroutes.
We traverse the nodes/settings/edges to compute data descriptions and when encountering a CPU node we attempt to look up the data registered in the compute graph. Reroutes are culled before compute graph compilation we need to transparently jump over them when computing data descriptions.

This issue in theory applies to any node that is statically culled during compilation, however i think reroutes are the only problem case. Branch culling should be a non-issue by definition and higen works out because upstream grid size nodes are present. If we statically cull more things in the future, we might want to make a deeper fix - like have a full mapping of input pins->upstream output pins so we have the graph connectivity after culling, but this seems like a good low-impact step and might be sufficient.

#jira UE-225579
#rnx
#rb julien.lheureux, wyatt.marvil

[CL 36761413 by huw bowles in 5.5 branch]