Fixed compilation non-determinism in animation blueprints
Name-generation code in UAnimGraphNode_Base::GetDescriptiveCompiledName is taking the hash of a node's GUID, expecting it to be stable, but the calling code was passing the duplicated node, not the original, so the GUID varied each time. This fix passes the 'source' node instead for stability. Any additional duplicates are then handled as before by the NetNameMap incrementing the name's number.

#jira UE-226025
#rb Jurre.deBaare
#rnx

[CL 36764471 by thomas sarkanen in 5.5 branch]