[Backout] - CL36699932
Causes issues in motion matching samples with parameters not updating correctly. Possibly due to blend stack graphs with same name.

[FYI] Thomas.Sarkanen
Original CL Desc
-----------------------------------------------------------------
Fixed compilation non-determinism in animation blueprints

Name-generation code in UAnimGraphNode_Base::GetDescriptiveCompiledName is taking the hash of a node's GUID, expecting it to be stable, but the calling code was passing the duplicated node, not the original, so the GUID varied each time. This fix passes the 'source' node instead for stability. Any additional duplicates are then handled as before by the NetNameMap incrementing the name's number.

#jira UE-226025
#rb Nicholas.Frechette, samuele.rigamonti
#rnx
#tests PIE, Anim BP Compile

[CL 36766127 by daren cheng in 5.5 branch]