[Backout] - CL36474486
Requested backout by Thomas due to CL36699932 backout.

[FYI] Thomas.Sarkanen
Original CL Desc
-----------------------------------------------------------------
Fixed rewiring animation blueprint nodes occasionally not picking up changes correctly

When rewiring nodes of the same type (say two sequence players each playing different animations), depending on the combination of default values and the current value in the running instance, re-instancing could potentially end up not giving the user what they expect (say, playing a different animation). This is because nodes are given numeric property names during compilation, and distinct nodes can be considered 'the same' depending on how nodes are rewired and processed. When it comes to reinstancing, the state of the node in the existing instance can end up 'winning', despite the node's info coming from a different source.

The solution to this is to uniquely name each node in an invariant way according to the source UEdGraphNode that they come from. In this case, to preserve name table size, I've chosen to use the hash of the node's GUID so the resulting FName can continue to just use a numeric extension rather than a whole new entry.

#jira UE-221734
[RN] Fixed rewiring animation blueprint nodes occasionally not picking up changes correctly
#rb Nicholas.Frechette
#tests PIE, Anim BP Compile

[CL 36766134 by daren cheng in 5.5 branch]