Fix for a GameplayGraph Test that was improperly initializing a struct.
The constructor does not allow brace initialization for structs if the struct's variables are default initialized in the header with its definition. FGraphProperties::bGenerateIslands is being initialized to true as part of the struct definition. Rather than change the struct definition, the brace initialization in this test was just removed since it was re-setting bGenerateIslands to its default value

#jira UE-224067
#rb ben.hoffman

[CL 36797012 by graham matuszewski in 5.5 branch]