Fix cooking issue with groom binding on certain platforms.
The skel. mesh DDC is not stable, and can mutate. This causes the indexing of the groom binding's cached cooked data to be incorrect. Ultimately, the cooking will fails due to missing cooked data for groom binding.
This CL forces the skel. mesh data to be cached when calling UGroomBindingAsset::BeginCacheForCookedPlatformData() prior to building the binding DDC key. This way, the Skel. mesh DDC key will be up to date and groom binding caching will be correctly indexed, and could be retrieved later when serializing the data. This change only impacts cooking (no impact for editor/runtime part).

#rb danny.couture

[CL 36827036 by charles derousiers in 5.5 branch]