[UE-226551] Fix Crash in Harmonix::Midi::Constants::SubdivisionToBeats()
No need to "assert" when passing in an invalid Quantization enum. We return "0" which can be handled gracefully.

However, handling this "gracefully" requires a bit of extra work. The PulseGenerator needs to not send (or stop sending) pulses when then Interval is invalid. Also, if the interval changes from valid to invalid, we need to stop active pulses.

Also, handle some edge cases if the Time Signature is invalid as well.

[REVIEW] [at]buzz.burrowes [at]charlie.huguenard

#jira UE-226551
#rb buzz.burrowes
#tests Harmonix Automation Tests & Editor Tests
#rnx

[CL 36880289 by jake burga in 5.5 branch]