[Water]
Removed the transient flag from the water brush capture/jumpflood components. The editor does not handle loading of non-transient-but-now-transient actor components very well for blueprint classes. In some _very_ rare cases, a previously serialized component will load and try to attach itself to the CDO of the blueprint rather than having the AttachParent property remapped to point to the instance itself.

#rb jonathan.bard

[CL 36898678 by roey borsteinas in 5.5 branch]