Sequencer: Don't disable the filter bar widget when muting because then you can't right click on it to unmute filters
Added SetMuted() to the filter bar widget which will disable the search box, horizontal and vertical containers. This context menu is still enabled so allows for the right click menu to process.

Also, make sure the muted/enabled state is synced when the filter bar widget is rebuilt (ie. switching layout from horizontal to vertical or toggling visibility)

#jira UE-226720
#rb Mike.Zyracki
#rnx

[CL 36930735 by max chen in 5.5 branch]