[ChaosVD] Fix for an issue where sphere collision shapes inside of a static mesh are visualized at the incorrect location.
There issue is caused because when we add a sphere collision that has an offset in its position relative to its parent transform, we just change the center position. CVD was only handling such scenarios for implicit transformed objects. The fix is just take into account the center location when we calculate the final transform, the same way as we do it for capsules already.

[REVIEW]
#jira UE-226827
#rb Benn.Gallagher

[CL 36950177 by sergio gardeazabal in 5.5 branch]