Fix cook of groom binding asset on certain target platform, due to skel. mesh updating their DDC when building their data.
The initial fix (CL 36826968) caused a deadlock as it involved taking a lock on the skel mesh, and was reverted in CL 36936042. Instead, this new CL recomputes the binding asset platform data a second time if the skel. mesh DDC key is not identical before and after the groom binding building.

#rb skip

[CL 37014011 by charles derousiers in 5.5 branch]