Taking a more belt and braces approach to NDC layout and proxy lifetimes to account for all edge cases.
UNiagaraDataChannel
Moved some layout information that is referenced by DataSets in handlers and DIs out to a utility struct that is alloacted and distributed as a shared ptr.
Avoids issue with lifetime management when the data channel object is destroyed but we still have in flight DIs and handlers.


FNiagaraDataChannelDataProxy
Converting these to be allocated and distributed as a shared ptr also.
This allows us to ensure it's lifetime in all cases with dispatcher/FXsystem/world distruction as well as NDC and DI destrution.

Potental issues this avoids -
1. NDC data is destroyed.
Both the dispatcher and DI Proxies hold refs to the NDCDataProxy.
We can submit an RT command to destroy the proxy and remove it from the dispatcher but cannot immediately and safely remove the dangling ptrs from potentially in flight DI proxies.
These will refresh next frame due to the GT side data having been destroyed but it leaves these dangling pointers temproarily in the DI Proxy and the potential for use.
I don't think we've hit this in practice yet but it's possible.

2. Simplifies and solidifies when the FScene changes and the dispatcher is destroyed.
UWorld cleanup flows can be janky. The FScene/Dispatcher on UWorlds can be swapped early etc.
We can't rely on getting a valid dispatcher ptr from the UWorld in all cases.
Using a shared ptr for the proxy allows the GT to just forget it's proxy and create a new one.
The old proxy can be destroyed from the dispatcher or the DI's RT proxy.
Any cleanup needing the dispatcher ptr can be done from the dispatcher itself which removes the previous possibility of mis-matched dispatchers when changing FScene.


The NDCDataProxy now also fully owns it's FNiagaraDataSet. Prevously this was partly managed by the GT side data which could lead to issues.
Now it's created along with the proxy but used and destroyed solely on the RT.
NDCDataProxy also keeps a reference to the layout info it needs from the main UNiagaraDataChannel.

FNiagaraDataBufferReadback

Changed helper class for reading back data from the CPU to be safer wrt data set lifetimes.
It's not safe to pass niagara buffers owned by the RT data set back to the GT.
Modified this class to pass the buffers back indivudially which are then coped over to a GT owned data buffer.



- Misc
Fixed FNiagaraDataChannelManager::RemoveDataChannel() to properly clean up the removed handler.
Island NDCs are clearing their NDCData when the islands are cleaned up. Currently recreating these anyway on acquire so this just holds onto memory longer than needed.

#rb Stu.McKenna

[CL 37045318 by simon tovey in 5.5 branch]