[ChaosVD] Fix for an issue where if async physics is enabled and the GT ticks at a lower delta time than the PT, CVD will discard game thread data even if it was properly recorded.
This happened due to an issue with the "auto trim" feature I implemented very early on, which pre-dates proper async physics support. We re-use the Frame Start/End event already available in UE trace and as we don't have control over them, this means we can have frames with no valid CVD data, and that is why we want to remove them from the recording.

The problem is that since that was implemented, now an "empty frame" is not the same as it was before. Previously an empty frame was a frame with no solver frame data associated with it, but nowadays game frames can have GT specific data like scene queries.

Another issue is that with async physics, the end of a GT frame can happen while a PT frame is also being processed, so we need to evaluate the frames we didn't commit to the CVD recording yet before determining a GT frame is indeed empty.

[REVIEW]
#jira UE-227009
#rb Benn.Gallagher

[CL 37062431 by sergio gardeazabal in 5.5 branch]