[ChaosVD] Fix for playback controls getting stuck while scrubbing in some situations
A few weeks ago we introduced support for "in-between" solver stages data, for any recorded data that happens outside a pre-defined solver stage. Examples of this are particle creation/destruction events.

That work required to auto-generate a solver stage structure on the fly in CVD when a CVD file is loaded. The problem is that the UI code assumed the number stages to be stable between frames, which is no longer true with this auto-generated solver state system.

This CL just make the UI pass INDEX_NONE as solver stage to the playback controller, so the controller can just play the last available stage in these situations. We are still checking and erroring out if we pass a specific stage number that is not valid.

#jira UE-227013
#rb Benn.Gallagher, Chris.Caulfield

[CL 37232150 by sergio gardeazabal in 5.5 branch]