UVEditor: Fixed crash by preventing UV Layer property from being reset to default. Fixed Snapshot tool being disabled when launching the UV Editor from the Modeling Tools palette
For the crash with the UV Layer property, we fix the crash by preventing ResetToDefault on the property. However, we also hide this property to bring it in line with the way other tools operate in the UV Editor. Users are expected to change the UV Channel from the UV Editor toolbar which will cancel tools that are already in flight on a different UV channel.

#rb tyson.brochu
#jira UE-227590

[CL 37289168 by lonnie li in 5.5 branch]